refactor: change hooks to normal structure
